Tidied up boid class
The boid class had gathered a lot of unnecessary stuff from previous iterations, so I have cleared all of that out. When I work on the zombie implementation of the boids, I will use a new class and use the boid one as a reference.

I have also added a slide for speed and screen turn factor.