Player chararacters detect zombies
Player characters can now detect zombies entering and exiting their areas and they can spawn an arrow that moves in the zombies' direction.
The arrow part is very rudimentory and needs more work. It was just a test.

I also need to see if the zombie leave the area if it is destroyed or if more work needs to be done to make sure it doesn't leave any null objects in the player's zombie array.