Player characters can shoot zombies
Player characters detect zombies in their range and remove them when out of their range.
When in range, they send their own node to the zombie so that, when the zombie dies, it can tell the player to remove itself from their zombie list.
The arrow removes the zombie from the game. It removes itself upon contact with anything.