Added health system to zombie
It's a rudimentary health system, but it is working.

I've made it it's own scene. It has a health variable, take damage function, and on death signal. The owner (zombie/player/etc) needs to connect to the signal and have a get healthsystem function so other's can access it.
I feel like there is a better way to implement it but this isn't the worst way. It can still be used by different entities and it's okay for small games. It would just be limited when scaling.