Created GameManager
The GameManager scene is in charge of running the game. It is a singleton and is the go to script for others to find the current state, such as paused, game over, etc.

In this update, the character and zombie spawners don't start until the GameManager emits the start game signal (currently controlled by pressing space.